Added tests for words

+++ app/routers/words.py
@@ -4,8 +4,9 @@ from app.database import get_db
 from sqlalchemy.orm import Session
 from app import models, schemas, oauth2
 from typing import List
+from indic_transliteration import sanscript
+from indic_transliteration.sanscript import transliterate
 from app.utils.lang import isDevanagariWord
-
 
 router = APIRouter(
     prefix="/words",
@@ -42,7 +43,7 @@ def get_words(db: Session = Depends(get_db)):
         
     return words
 
-@router.get("/{id}")
+@router.get("/{id}", response_model=schemas.WordOut)
 def get_word(id: int, db: Session = Depends(get_db)):
     db_word = db.query(models.SanskritWord).filter(models.SanskritWord.id == id).first()
 
@@ -72,157 +73,147 @@ def get_word(id: int, db: Session = Depends(get_db)):
 
 @router.post("/")
 def create_word(word: schemas.Word, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
-    if current_user.is_superuser == False:
-        raise HTTPException(status_code=status.HTTP_403_FORBIDDEN, detail="Not authorized to perform requested action")
-    
-
     db_word = db.query(models.SanskritWord).filter(models.SanskritWord.sanskrit_word == word.sanskrit_word).first()
 
     if db_word:
-        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="Word already exists")
+        raise HTTPException(status_code=status.HTTP_409_CONFLICT, detail="Word already exists")
+
+    if not isDevanagariWord(word.sanskrit_word):
+        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="Word is not in Devanagari")
     
     new_word = models.SanskritWord(
         sanskrit_word = word.sanskrit_word,
-        english_word = "hi"
+        english_word = transliterate(word.sanskrit_word, sanscript.DEVANAGARI, sanscript.HK),
     )
     db.add(new_word)
     db.flush()
 
     if word.etymology:
-        for etymology in word.etymology:
-            db.add(models.Etymology(
-                sanskrit_word_id = new_word.id,
-                etymology = etymology
-            ))
-    
+        etymologies = [models.Etymology(
+            sanskrit_word_id = new_word.id,
+            etymology = etymology
+        ) for etymology in word.etymology]
+        db.add_all(etymologies)
+
     if word.derivation:
-        for derivation in word.derivation:
-            db.add(models.Derivation(
-                sanskrit_word_id = new_word.id,
-                derivation = derivation
-            ))
+        derivations = [models.Derivation(
+            sanskrit_word_id = new_word.id,
+            derivation = derivation
+        ) for derivation in word.derivation]
+        db.add_all(derivations)
 
 
     if word.translation:
-        for translation in word.translation:
-            db.add(models.Translation(
-                sanskrit_word_id = new_word.id,
-                english_translation = translation.english_translation,
-                kannada_translation = translation.kannada_translation,
-                hindi_translation = translation.hindi_translation,
-                detailedDescription = translation.detailedDescription,
-            ))
+        translations = [models.Translation(
+            sanskrit_word_id = new_word.id,
+            english_translation = translation.english_translation,
+            kannada_translation = translation.kannada_translation,
+            hindi_translation = translation.hindi_translation,
+            detailedDescription = translation.detailedDescription,
+        ) for translation in word.translation]
+        db.add_all(translations)
     
     
     if word.reference_nyaya_text:
-        for reference in word.reference_nyaya_text:
-            db.add(models.ReferenceNyayaText(
-                sanskrit_word_id = new_word.id,
-                source = reference.source,
-                description = reference.description
-            ))
+        references = [models.ReferenceNyayaText(
+            sanskrit_word_id = new_word.id,
+            source = reference.source,
+            description = reference.description
+        ) for reference in word.reference_nyaya_text]
+        db.add_all(references)
     
     
     if word.synonyms:
-        for synonym in word.synonyms:
-            db.add(models.Synonym(
-                sanskrit_word_id = new_word.id,
-                synonym = synonym
-            ))
+        synonyms = [models.Synonym(
+            sanskrit_word_id = new_word.id,
+            synonym = synonym
+        ) for synonym in word.synonyms]
+        db.add_all(synonyms)
     
     
     if word.antonyms:
-        for antonym in word.antonyms:
-            db.add(models.Antonym(
-                sanskrit_word_id = new_word.id,
-                antonym = antonym
-            ))
-
+        antonyms = [models.Antonym(
+            sanskrit_word_id = new_word.id,
+            antonym = antonym
+        ) for antonym in word.antonyms]
+        db.add_all(antonyms)
+    
     db.commit()
 
     return JSONResponse(status_code=status.HTTP_201_CREATED, content={"message": "Word created"})
 
 @router.put("/")
-def update_word(word: schemas.WordOut, db: Session = Depends(get_db)):
+def update_word(word: schemas.WordOut, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
     db_word = db.query(models.SanskritWord).filter(models.SanskritWord.id == word.id).first()
 
     if not db_word:
         raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Word with id {word.id} not found")
     
+    
     db_word.sanskrit_word = word.sanskrit_word
     db_word.english_word = word.english_word
 
     if word.etymology:
-        ety = db.query(models.Etymology).filter(models.Etymology.sanskrit_word_id == word.id).first()
-        if ety:
-            ety.etymology = word.etymology
-        else:
-            db.add(models.Etymology(
-                sanskrit_word_id = word.id,
-                etymology = word.etymology
-            ))
+        db.query(models.Etymology).filter(models.Etymology.sanskrit_word_id == word.id).delete()
+
+        etymologies = [models.Etymology(
+            sanskrit_word_id = word.id,
+            etymology = etymology
+        ) for etymology in word.etymology]
         
+        db.add_all(etymologies)
+
     if word.derivation:
-        der = db.query(models.Derivation).filter(models.Derivation.sanskrit_word_id == word.id).first()
-        if der:
-            der.derivation = word.derivation
-        else:
-            db.add(models.Derivation(
-                sanskrit_word_id = word.id,
-                derivation = word.derivation
-            ))
-    
+        db.query(models.Derivation).filter(models.Derivation.sanskrit_word_id == word.id).delete()
+
+        derivations = [models.Derivation(
+            sanskrit_word_id = word.id,
+            derivation = derivation
+        ) for derivation in word.derivation]
+        
+        db.add_all(derivations)
+
 
     if word.translation:
-        trans = db.query(models.Translation).filter(models.Translation.sanskrit_word_id == word.id).first()
-        if trans:
-            trans.english_translation = word.english_translation
-            trans.kannada_translation = word.kannada_translation
-            trans.hindi_translation = word.hindi_translation
-            trans.detailedDescription = word.detailedDescription
-        else:
-            db.add(models.Translation(
-                sanskrit_word_id = word.id,
-                english_translation = word.english_translation,
-                kannada_translation = word.kannada_translation,
-                hindi_translation = word.hindi_translation,
-                detailedDescription = word.detailedDescription
-            ))
-    
+        db.query(models.Translation).filter(models.Translation.sanskrit_word_id == word.id).delete()
 
+        translations = [models.Translation(
+            sanskrit_word_id = word.id,
+            english_translation = translation.english_translation,
+            kannada_translation = translation.kannada_translation,
+            hindi_translation = translation.hindi_translation,
+            detailedDescription = translation.detailedDescription,
+        ) for translation in word.translation]
+        
+        db.add_all(translations)
+    
+    
     if word.reference_nyaya_text:
-        ref = db.query(models.ReferenceNyayaText).filter(models.ReferenceNyayaText.sanskrit_word_id == word.id).first()
-        if ref:
-            ref.source = word.source
-            ref.description = word.description
-        else:
-            db.add(models.ReferenceNyayaText(
-                sanskrit_word_id = word.id,
-                source = word.source,
-                description = word.description
-            ))
+        db.query(models.ReferenceNyayaText).filter(models.ReferenceNyayaText.sanskrit_word_id == word.id).delete()
+
+        references = [models.ReferenceNyayaText(
+            sanskrit_word_id = word.id,
+            source = reference.source,
+            description = reference.description
+        ) for reference in word.reference_nyaya_text]
+        
+        db.add_all(references)
     
-
+    
     if word.synonyms:
-        syn = db.query(models.Synonym).filter(models.Synonym.sanskrit_word_id == word.id).first()
-        if syn:
-            syn.synonym = word.synonyms
-        else:
-            db.add(models.Synonym(
-                sanskrit_word_id = word.id,
-                synonym = word.synonyms
-            ))
-
+        synonyms = [models.Synonym(
+            sanskrit_word_id = word.id,
+            synonym = synonym
+        ) for synonym in word.synonyms]
+        db.add_all(synonyms)
+    
     
     if word.antonyms:
-        ant = db.query(models.Antonym).filter(models.Antonym.sanskrit_word_id == word.id).first()
-        if ant:
-            ant.antonym = word.antonyms
-        else:
-            db.add(models.Antonym(
-                sanskrit_word_id = word.id,
-                antonym = word.antonyms
-            ))
+        antonyms = [models.Antonym(
+            sanskrit_word_id = word.id,
+            antonym = antonym
+        ) for antonym in word.antonyms]
+        db.add_all(antonyms)
 
     db.commit()
     db.refresh(db_word)
@@ -231,7 +222,12 @@ def update_word(word: schemas.WordOut, db: Session = Depends(get_db)):
 
 
 @router.delete("/{id}")
-def delete_word(id: int, db: Session = Depends(get_db)):
+def delete_word(id: int, db: Session = Depends(get_db), current_user: int = Depends(oauth2.get_current_user)):
+    db_word = db.query(models.SanskritWord).filter(models.SanskritWord.id == id).first()
+
+    if not db_word:
+        raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail=f"Word with id {id} not found")
+
     db.query(models.SanskritWord).filter(models.SanskritWord.id == id).delete()
     db.query(models.Etymology).filter(models.Etymology.sanskrit_word_id == id).delete()
     db.query(models.Derivation).filter(models.Derivation.sanskrit_word_id == id).delete()
@@ -242,4 +238,4 @@ def delete_word(id: int, db: Session = Depends(get_db)):
 
     db.commit()
 
-    return JSONResponse(status_code=status.HTTP_200_OK, content={"message": "Word deleted"})
+    return JSONResponse(status_code=status.HTTP_204_NO_CONTENT, content={"message": "Word deleted"})
